Add dagger hilt Move activity and fragments to new ui package
