+++ app/src/main/java/hector/ruiz/rickandmorty/App.kt
@@ -1,0 +1,7 @@
+package hector.ruiz.rickandmorty
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
