+++ app/src/main/java/hector/ruiz/rickandmorty/ui/DetailFragment.kt
@@ -1,4 +1,4 @@
-package hector.ruiz.rickandmorty
+package hector.ruiz.rickandmorty.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
