+++ app/src/main/res/layout/detail_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DetailFragment">
+    tools:context=".ui.DetailFragment">
 
     <TextView
         android:id="@+id/textview_second"
