+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/ListFragment"
-        android:name="hector.ruiz.rickandmorty.ListFragment"
+        android:name="hector.ruiz.rickandmorty.ui.ListFragment"
         android:label="@string/list_fragment_label"
         tools:layout="@layout/list_fragment">
 
@@ -18,7 +18,7 @@
 
     <fragment
         android:id="@+id/DetailFragment"
-        android:name="hector.ruiz.rickandmorty.DetailFragment"
+        android:name="hector.ruiz.rickandmorty.ui.DetailFragment"
         android:label="@string/detail_fragment_label"
         tools:layout="@layout/detail_fragment">
