+++ build.gradle
@@ -10,6 +10,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:$rootProject.gradlePluginVersion"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$rootProject.kotlinPluginVersion"
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$rootProject.daggerHiltVersion"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
