+++ gradle/versions.gradle
@@ -28,6 +28,9 @@ ext {
     // Moshi
     moshiVersion = "1.12.0"
 
+    // Dagger Hilt
+    daggerHiltVersion = "2.38.1"
+
     // Testing Libraries
     junitVersion = "4.13.2"
     testJunitVersion = "1.1.3"
